author cannot delete or show or edit others post

+++ app/Http/Controllers/Author/PostController.php
@@ -89,20 +89,27 @@ class PostController extends Controller
         return redirect()->route('author.post.index');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Post  $post
-     * @return \Illuminate\Http\Response
-     */
+
+
     public function show(Post $post)
     {
+        if($post->user_id != auth()->id())
+        {
+            Toastr::error('You are not authorized to access this post!', 'Error');
+            return redirect()->back();
+        }
         return  view('author.post.show', compact('post'));
     }
 
 
     public function edit(Post $post)
     {
+        if($post->user_id != auth()->id())
+        {
+            Toastr::error('You are not authorized to access this post!', 'Error');
+            return redirect()->back();
+        }
+
         $categories = Category::all();
         $tags = Tag::all();
         return view('author.post.edit', compact('post','categories', 'tags'));
@@ -174,6 +181,13 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
+
+        if($post->user_id != auth()->id())
+        {
+            Toastr::error('You are not authorized to access this post!', 'Error');
+            return redirect()->back();
+        }
+
         if ( Storage::disk('public')->exists('post/'.$post->image) ){
             Storage::disk('public')->delete('post/'.$post->image);
         }
